Removed reduandant code and made code use arraylists instead of linkedlists for better performance

+++ src/BoardPiece.java
@@ -1,17 +1,17 @@
-import java.util.LinkedList;
+import java.util.ArrayList;
 
 public class BoardPiece {
     public int xPos, yPos;
     public int scale;
     public boolean isWhite;
-    public LinkedList<BoardPiece> pieces;
+    public ArrayList<BoardPiece> pieces;
     public int x, y;
     public boolean hasMoved;
 
-    public BoardPiece(int xPos, int yPos, boolean isWhite, LinkedList<BoardPiece> pieces, int scale) {
+    public BoardPiece(int xPos, int yPos, boolean isWhite, ArrayList<BoardPiece> pieces, int scale) {
         this.xPos = xPos;
         this.yPos = yPos;
-        this.scale = scale;
+        this.scale = Main.size;
         x = xPos * scale;
         y = yPos * scale;
         this.isWhite = isWhite;
@@ -22,21 +22,15 @@ public class BoardPiece {
 
     public void move(int xPos, int yPos) {
         BoardPiece piece = Main.getPiece(xPos * scale, yPos * scale);
-        LinkedList<Vector2d> legalSquares = getLegalSquares();
+        ArrayList<Vector2d> legalSquares = getLegalSquares();
         Vector2d testVector = new Vector2d(xPos, yPos);
         if (!legalSquares.contains(testVector) || !(Main.isWhiteTurn == isWhite)) {
             x = this.xPos * scale;
             y = this.yPos * scale;
             return;
         }
-        if (piece != null) {
-            if (piece.isWhite != isWhite) {
-                piece.kill();
-            } else {
-                x = this.xPos * scale;
-                y = this.yPos * scale;
-                return;
-            }
+        if (piece != null && piece.isWhite != isWhite) {
+            piece.kill();
         }
         this.xPos = xPos;
         this.yPos = yPos;
@@ -50,11 +44,11 @@ public class BoardPiece {
         pieces.remove(this);
     }
 
-    public LinkedList<Vector2d> getLegalSquares() {
+    public ArrayList<Vector2d> getLegalSquares() {
         return null;
     }
 
-    public void printMoves(LinkedList<Vector2d> moves) {
+    public void printMoves(ArrayList<Vector2d> moves) {
         for (Vector2d move : moves) {
             System.out.print(move.x + " " + move.y + " ");
         }
